Refactor out markdown parsing
